Move EditExpenseModal to its own file in components/
